+++ web/src/pages/auth/login/LoginForm.module.css
@@ -1,0 +1,9 @@
+.oauthButtonWrapper {
+  display: flex;
+  flex-direction: column;
+  gap: 1.25rem;
+
+  @media (min-width: 640px) {
+    flex-direction: row;
+  }
+}
